Migrate team info to competition wrapper
